LPD-93373 Move test to the proper test class
